+++ 斐波纳契序列递归法超时.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 //定义函数递归法计算斐波纳契序列
-int toSolve(int n){ 
+long long toSolve(int n){ 
     if (n==0)
     {
         return 1;   
@@ -23,7 +23,7 @@ int main(int argc, char const *argv[])
     scanf("%d",&n);
 
     //输出
-    printf("%d",toSolve(n));
+    printf("%lld",toSolve(n));
     return 0;
 }
